Refactor: directory structure
