+++ src/components/organisms/category/index.js
@@ -1,5 +1,4 @@
 export { default as Advertisement } from './Advertisement';
-export { default as Board } from './Board';
-export { default as CategoryList } from './CategoryList';
+export { default as Board } from '../../pages/articles/ArticleList/Board';
 export { default as LeftBar } from './LeftBar';
-export { default as PageSelector } from './PageSelector';
+export { default as PageSelector } from '../../../common/PageSelector';
